refactor: update import statements for new type structure
- Update all existing files to import from new type files (entities.ts, inputs.ts, shared.ts)
- Replace imports from obsolete src/types/data.ts
- Maintain type safety and compatibility across the entire codebase
- Remove legacy type definition file

This completes the type definition restructuring for improved maintainability and separation of concerns.
