+++ src/components/QuestionsSection.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 import { describe, expect, test, vi } from 'vitest';
 import { QuestionsSection } from './QuestionsSection';
 import { renderWithProviders, testUtils } from '../test/test-utils';
-import type { Question } from '../types/data';
+import type { Question } from '../types/entities';
 
 // テスト用のモックデータ
 const mockQuestions: Question[] = [
