+++ src/hooks/useNurseryEdit.test.ts
@@ -6,7 +6,7 @@
 import { renderHook, act } from '@testing-library/react';
 import { vi } from 'vitest';
 import { useNurseryEdit } from './useNurseryEdit';
-import type { Nursery } from '../types/data';
+import type { Nursery } from '../types/entities';
 
 // モック関数の作成
 const mockUpdateNursery = vi.fn();
